Get rid of splint error. Proper dependencies for the example programs.

+++ drill/drill_util.c
@@ -13,7 +13,7 @@
 
 #include <errno.h>
 
-static int
+static size_t
 read_line(FILE *input, char *line, size_t len)
 {
 	size_t i;
@@ -52,7 +52,7 @@ read_key_file(const char *filename, ldns_rr_list *key_list)
 		return LDNS_STATUS_ERR;
 	}
 	while (line_len >= 0) {
-		line_len = read_line(input_file, line, sizeof(line));
+		line_len = (int) read_line(input_file, line, sizeof(line));
 		line_nr++;
 		if (line_len > 0 && line[0] != ';') {
 			status = ldns_rr_new_frm_str(&rr, line, 0, NULL, NULL);
